initial implementation of the finite state machine
This change introduces a simple implementation of a finite state machine. This initial implementation does not have support for events or callbacks.

This initial implementation simply tracks machine state, and whether a transition from state A to state B is permitted. Future improvements will add the ability to add callbacks and use events to change the state of the machine.